+++ app/components/RoleSelector/index.tsx
@@ -4,70 +4,16 @@ import { UserContext } from "@/app/contexts/UserContext";
 import { Role } from "@/app/types";
 import { getByCookies } from "@/app/utils/https";
 import { useContext, useEffect, useState } from "react";
-import { RoleAvatar, UserAvatar } from "@/app/components/Avatar";
-import Button from "@/app/components/Button";
-import { useRouter } from "next/navigation";
+import { RoleAvatar, UserAvatar } from "@/app/components/RoleSelector/Avatar";
 import { Tooltip } from "flowbite-react";
 import ProfileCard from "@/app/components/ProfileCard";
+import { AddRole } from "./AddRole";
+import { More } from "./More";
 
 interface RoleSwitcherProps {
   selectedRole: Role | null,
   onSelectedRole?: (role: Role) => void;
 }
-
-const AddRole = () => {
-  const router = useRouter();
-
-  const handleCreateRole = () => {
-    router.push("/settings/create-role");
-  };
-
-  return (
-    <Button
-      icon={{
-        img: {
-          src: "/AddIconImage.jpg",
-          alt: "Add Role"
-        },
-        hoverable: false,
-      }}
-      className="opacity-50 hover:opacity-70 hover:bg-black"
-      onClick={handleCreateRole}
-    />
-  );
-};
-
-const More = () => {
-  return (
-    <Button
-      icon={{}}
-      className="px-3"
-    >
-      <div className="flex flex-col items-center text-xs text-gray-300 hover:text-gray-200">
-        <div>
-          <svg
-            className="w-8 h-8"
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            fill="none"
-            viewBox="0 0 24 24"
-          >
-            <path
-              stroke="currentColor"
-              strokeLinecap="round"
-              strokeWidth="2"
-              d="M6 12h.01m6 0h.01m5.99 0h.01"
-            />
-          </svg>
-        </div>
-        <div>
-          More
-        </div>
-      </div>
-    </Button>
-  );
-};
 
 export default function RoleSelector(props: RoleSwitcherProps) {
   const { user } = useContext(UserContext);
